Add Design spec [incomplete]
The initial design specs contains detailed information about what the project's purpose is, what it does, who will use it, what is the problem it is trying to solve and why does it need to exist. Other information contained within this word document is some initial requirements of the project and a high level entity diagram of the main structures. Before the diagram of high level entities, there is a break down of what each structure of the project is aimed to achieve.